lambda/requirements.txt
aboto3==1.34.0
azure-storage-blob==12.19.1